requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
click==8.1.8
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
fastapi==0.115.6
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
h11==0.14.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
motor==3.6.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
strawberry-graphql==0.258.0
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
wcwidth==0.2.13